feat!: migrate from Redis to SQLite
BREAKING CHANGE: all features that relied on Redis to store data have
been migrated to use SQLite. A migration script is provided and it can
be run with `bun src/scripts/redis_to_sqlite.ts`. It will transfer all
data from Redis to a SQLite database.

REDIS_PORT can be set if your Redis instance is not listening on port
6379.
DATABASE_PATH can be set to change the location of the SQLite database,
it default to `bot.sqlite` in the current directory.

The functionality provided by Redisearch to perform full text searches
on quotes is now provided by the fts5 SQLite extension (https://www.sqlite.org/fts5.html).
Since not all SQLite installations include the extension, it is possible
to load it dynamically at runtime by providing the path to the extension
in the SQLITE_EXTENSIONS environment variable.
Example: SQLITE_EXTENSIONS=/path/to/fts5.so

In a future version of this bot, the Redis client library and the
migration script will be removed.